fix warnings during building (#567)
* make some functions from mdb_iamproxy.c static

* fix new line symbol missing

* make apply_fmt

* fix warnings during building

* make apply_fmt

* fix issues

---------

+++ sources/frontend.c
@@ -1561,8 +1561,9 @@ static inline od_frontend_status_t od_frontend_poll_catchup(od_client_t *client,
 	int absent_heartbeat_checks = 0;
 	while (route->last_heartbeat == 0) {
 		machine_sleep(ODYSSEY_CATCHUP_RECHECK_INTERVAL);
-		if (absent_heartbeat_checks++ >
-		    (timeout * 1000 / ODYSSEY_CATCHUP_RECHECK_INTERVAL)) {
+		if ((int64_t)absent_heartbeat_checks++ > // add cast to int64_t for correct camparison (int64_t > int and ibt64_t > uint32_t)
+		    (int64_t)(timeout * 1000 /
+			      ODYSSEY_CATCHUP_RECHECK_INTERVAL)) {
 			od_debug(&instance->logger, "catchup", client, NULL,
 				 "No heartbeat for route detected\n");
 			return OD_ECATCHUP_TIMEOUT;
